Do not test linux-specific cases on OSX in dockertools
